[Fix] error mesg

+++ verbiverse/CustomWidgets/CFileListWidget.py
@@ -22,8 +22,9 @@ class CFileListWidget(QWidget, Ui_CFileListWidget):
         signalBus.open_localfile_signal.connect(self.addFile)
 
     def updateConfig(self):
-        with open(CONFIG_PATH + "historyfilelist.json", "w", encoding="utf-8") as f:
-            json.dump(self.file_list, f)
+        if os.path.exists(CONFIG_PATH + "historyfilelist.json"):
+            with open(CONFIG_PATH + "historyfilelist.json", "w", encoding="utf-8") as f:
+                json.dump(self.file_list, f)
 
     def loadConfigFile(self):
         if os.path.exists(CONFIG_PATH + "historyfilelist.json"):
